add contidiional logic to avoid accidentally delete all datas.

+++ api/services/ShippingService.js
@@ -18,16 +18,19 @@ module.exports = {
   saveAll: async (shippings) => {
     // console.log('=== 1 new shippings ==>\n',shippings);
     try {
+      // avoid accidentally delete all datas.
+      if(shippings == undefined || shippings.length < 1)
+        return { msg:'運費資料有誤！', success: false };
       // clean up
       let findAll = await db.Shipping.findAll();
       let deleteAll = await* findAll.map((shipping) => {
-        shipping.destroy();
+        return shipping.destroy();
         // console.log('=== now this id is destroied ==>',shipping.id);
       });
       // save them all!
       let newShippings = shippings;
-      let savedShippings = await* shippings.map((shipping) => {
-        db.Shipping.create(shipping)
+      let savedShippings = await* newShippings.map((shipping) => {
+        return db.Shipping.create(shipping)
         // console.log('=== now this should created ==>\n',shipping);
       });
       // console.log('=== 2 now savedShippings length ==>', savedShippings.length);
